pyproject.toml
[project]
name = "final-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "invoke>=2.2.0",
    "matplotlib>=3.10.1",
    "numpy>=1.26.0,<2.2.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "tensorflow>=2.18.0,<2.19.0",
    "uvicorn[standard]>=0.34.0",
]
